implemented mulitple pages using BrowserRouter
