+++ package.json
@@ -16,6 +16,8 @@
     "electron-is-dev": "^2.0.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
+    "react-router": "6.4.2",
+    "react-router-dom": "^6.4.2",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.6",
     "wait-on": "^6.0.1",
